Add download_audio code

+++ video/download_video.py
@@ -1,4 +1,3 @@
-#Code Contributor - Ankit Shah - ankit.tronix@gmail.com
 import pafy
 import time
 import datetime
@@ -11,7 +10,7 @@ from tqdm import tqdm
 
 #Format video - Convert to H.264 and trim
 def format_video(input_video_file, output_video_file, start_time, duration):
-	cmdstring = "ffmpeg -loglevel panic -i {0} -ss 00:00:{1} -t 00:00:{2} {3}".format(input_video_file, start_time, duration, output_video_file)
+	cmdstring = "ffmpeg -loglevel panic -i {0} -ss {1} -t {2} {3}".format(input_video_file, start_time.split(".")[0], duration, output_video_file)
 	print(cmdstring)
 	os.system(cmdstring)
 
@@ -23,11 +22,9 @@ def download_video_method(line,csv_file):
 	query_id = line.split(",")[0];
 	start_seconds = line.split(",")[1];
 	end_seconds = line.split(",")[2];
-	video_duration = float(end_seconds) - float(start_seconds)
+	video_duration = float(end_seconds.split(".")[0]) - float(start_seconds.split(".")[0])
 	#positive_labels = ','.join(line.split(",")[3:]);
 	print("Query -> " + query_id)
-	
-	url = "https://www.youtube.com/watch?v=" + query_id
 
 	segmented_folder = sys.argv[1].split('.csv')[0].split("/")[-1] + "_" + csv_file.split('.csv')[0] +  "_" + "video_formatted_and_segmented_downloads"
 		
@@ -38,8 +35,8 @@ def download_video_method(line,csv_file):
 
 	if not os.path.isfile(path_to_segmented_video):
 		try:
-			video = pafy.new(url)
-			bestvideo = video.getbestvideo()
+			video = pafy.new(query_id)
+			bestvideo = video.getbestvideo() # Maybe some videos don't have streams without audio? so can try downloading with audio
 
 			#.csv - split - to get the folder information. As path is also passed for the video - creating the directory from the path where this video script is present. THus using second split to get the folder name where output files shall be downloaded
 			output_folder = sys.argv[1].split('.csv')[0].split("/")[-1] + "_" + csv_file.split('.csv')[0] +  "_" + "video_downloaded"
@@ -49,24 +46,25 @@ def download_video_method(line,csv_file):
 				os.makedirs(output_folder)
 			
 			path_to_download = output_folder + "/Y" + query_id + "." + bestvideo.extension
-			bestvideo.download(path_to_download)			
+			bestvideo.download(path_to_download, quiet=True)			
 				
 			format_video(path_to_download, path_to_segmented_video, start_seconds, video_duration)
 
-			#Remove the original video 
-			delete_original_file_cmd = "rm -f {0}".format(path_to_download)
-			os.system(delete_original_file_cmd)
+			if os.path.isfile(path_to_segmented_video):
+				#Remove the original video 
+				delete_original_file_cmd = "rm -f {0}".format(path_to_download)
+				os.system(delete_original_file_cmd)
+			else:
+				print("Error converting file: " + path_to_download)
 
 			ex1 = ""
 		except Exception as ex:
 			ex1 = str(ex) + ',' + str(query_id)
-			print ("Error is ---> " + str(ex))
+			print ("Error is ---> " + str(ex))  
 		return ex1
 
 
 def download_video(csv_file, timestamp):	
-	print(multiprocessing.cpu_count())
-
 	segments_info = []
 	
 	with open(csv_file, "r") as segments_info_file:
